Working on getting things converted to Variable types. Also working on the error propagations. Moved all of the gp and bash scripts into their own directory.
Makefile:
This has gotten some of the guys removed until I can get things fully
implemented as Variables. Only the eff recipe works.

EffCalculator.hpp/.cpp:
This has changes to make it coincide with the new ErrorCalculator class.
It also now handles things in terms of the Variable class.

ErrorCalculator.hpp/.cpp:
This is the new class that will calculate the errors on all of the various
Variables in the analysis. A single class to make all the calculations.

FileChecker.hpp/.cpp:
This is currently the error checking to see if the files/directories
that the program tries to use exist. This will probably need to be
rethought later.

SimConvoluter.hpp/.cpp:
This guy does the convolution of the simulation with the beta resolution.
It only needs to be done once for the analysis.

eff.cpp:
This was rewritten to take into account the new Variable class.


Signed-off-by: S.V. Paulauskas <stanpaulauskas@gmail.com>