Sensitivity calculation is mostly finished, adding theory scripts
Added the error bars on the scripts for plotting the densities. The lines
needed for the plots are also now output in ex.cpp.

I have added the scritps necessary to send to Michael for the CGM
calculations (GSimp1.dat and runInterface.bash). The most important thing
is to not have any blanks at the end of the file. This will cause errors
in the CGM code.

I have also set things up so that the calculations for the
parameterized variables in the fit are in their own class. This lets
them be accessable by everybody.

Signed-off-by: S.V. Paulauskas <stanpaulauskas@gmail.com>