Major changes, config file nearly implemented
Config.xml:
The configuration file for the program. Reads in at runtime.

Makefile:
Added some new stuff to the compilation

Configuration.cpp/.hpp:
This is the class that handles the parsing of all the configuration files
for the program. This is still being worked on at the moment.

Decay.cpp/.hpp:
This has been updated so that you now have all the various sets for the
different aspects of the decay. **NOTE** Still need to implement the
different pieces for the gating gammas.

Experiment.hpp:
Handles information about the experiment.

FileHandler.hpp/.cpp:
Handles all the file names/paths for the program.

FitHandler.hpp/.cpp:
Handles the information necessary for the fitting classes

FlagHandler.hpp/.cpp:
Handles all of the various flags for the different functions of the
program.

cfgtst.cpp:
This is the file used in the testing of the configuration class.

Signed-off-by: S.V. Paulauskas <stanpaulauskas@gmail.com>