Crystal Ball Variable Parameterization is now implemented
The parameterization for the Crystal Ball parameters with respect to the Time of Flight is now implemented through the configuration file instead of being hard coded in the TofFitter and other places. We still have some hard coded pieces to deal with in the CrystalBallParameters::CalcXXX functions since I do not now how to handle it at this time. We could also add some more robustness to match the names of the variables with the order that they are used.